requirements.txt
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
openai==1.12.0
python-dotenv==1.0.0
